backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
duckdb==1.1.3
orjson==3.10.12

# Auth (JWT/JWKS)
PyJWT[crypto]==2.9.0

# Optional AI provider (Conversation Service)
openai==1.59.6


